# docs/10_api_openapi.yaml
openapi: 3.0.3
info:
  title: Freight Forwarding and Finance Platform API
  version: 2.0.0
  description: >-
    Organization- and branch-scoped API for freight operations, informational
    service charges, reusable financial documents, and double-entry accounting.
servers:
  - url: /api/v1
    description: Current API
security:
  - bearerAuth: []
...
paths:
...
  /service-orders:
    get:
      tags: [Service Orders]
      operationId: listServiceOrders
      parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PageSize'}, {$ref: '#/components/parameters/BranchFilter'}, {name: status, in: query, schema: {type: string}}, {name: customer_party_id, in: query, schema: {type: integer}}]
      responses: {'200': {description: Service orders, content: {application/json: {schema: {$ref: '#/components/schemas/PagedServiceOrderResponse'}}}}}
  /service-orders/{serviceOrderId}:
    get:
      tags: [Service Orders]
      operationId: getServiceOrder
      parameters: [{$ref: '#/components/parameters/ServiceOrderId'}]
      responses: {'200': {description: Service order, content: {application/json: {schema: {$ref: '#/components/schemas/ServiceOrder'}}}}}
  /service-orders/{serviceOrderId}/containers:
    post:
      tags: [Service Orders]
      operationId: addActualContainer
      parameters: [{$ref: '#/components/parameters/ServiceOrderId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ActualContainerCreate'}}}
      responses: {'201': {description: Container created}}
  /service-orders/{serviceOrderId}/components:
    post:
      tags: [Service Orders]
      operationId: addServiceComponent
      parameters: [{$ref: '#/components/parameters/ServiceOrderId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ComponentCreate'}}}
      responses: {'201': {description: Component created}}
  /service-order-components/{componentId}/values:
    put:
      tags: [Service Orders]
      operationId: replaceComponentValues
      parameters: [{$ref: '#/components/parameters/ComponentId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ComponentValuesReplace'}}}
      responses: {'200': {description: Values saved}}
  /service-order-components/{componentId}/complete:
    post:
      tags: [Service Orders]
      operationId: completeComponent
      parameters: [{$ref: '#/components/parameters/ComponentId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'200': {description: Component completed}}
  /service-orders/{serviceOrderId}/charges:
    get:
      tags: [Service Charges]
      operationId: listServiceCharges
      parameters: [{$ref: '#/components/parameters/ServiceOrderId'}]
      responses: {'200': {description: Charges}}
    post:
      tags: [Service Charges]
      operationId: createServiceCharge
      parameters: [{$ref: '#/components/parameters/ServiceOrderId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ServiceChargeCreate'}}}
      responses: {'201': {description: Charge created}}
  /service-charges/{chargeId}/issue:
    post:
      tags: [Service Charges]
      operationId: issueServiceCharge
      parameters: [{$ref: '#/components/parameters/ChargeId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'200': {description: Charge issued; no accounting journal is created}}
  /service-charges/{chargeId}/create-finance-invoice:
    post:
      tags: [Service Charges]
      operationId: createFinanceInvoiceFromCharge
      parameters: [{$ref: '#/components/parameters/ChargeId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'201': {description: Draft financial invoice created}}
...
  /attachments/presign:
    post:
      tags: [Attachments]
      operationId: presignAttachment
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/AttachmentPresignRequest'}}}
      responses: {'200': {description: Presigned upload returned}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
...
    ServiceOrderId: {name: serviceOrderId, in: path, required: true, schema: {type: integer, format: int64}}
    ComponentId: {name: componentId, in: path, required: true, schema: {type: integer, format: int64}}
    ChargeId: {name: chargeId, in: path, required: true, schema: {type: integer, format: int64}}
...
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: page_size, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
    BranchFilter: {name: branch_id, in: query, schema: {type: integer, format: int64}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8, maxLength: 128}}
...
  schemas:
...
    PageMeta:
      type: object
      properties: {page: {type: integer}, page_size: {type: integer}, total: {type: integer}}
...
    ActualContainerCreate:
      type: object
      required: [container_type_id, container_number]
      properties: {container_requirement_id: {type: integer}, container_type_id: {type: integer}, container_number: {type: string}, seal_serial: {type: string}, net_weight_kg: {type: number}, gross_weight_kg: {type: number}}
    ComponentCreate:
      type: object
      required: [trade_direction_component_id]
      properties: {trade_direction_component_id: {type: integer}, sequence_no: {type: integer, minimum: 1}}
    ComponentValuesReplace:
      type: object
      required: [values]
      properties: {values: {type: array, items: {type: object, required: [template_attribute_id], properties: {template_attribute_id: {type: integer}, value_text: {type: string}, value_number: {type: number}, value_date: {type: string, format: date}, value_datetime: {type: string, format: date-time}, value_boolean: {type: boolean}, value_reference_type: {type: string}, value_reference_id: {type: integer}, value_json: {type: object}}}}}
    ServiceChargeCreate:
      type: object
      required: [document_type, document_date, currency_code, lines]
      properties: {document_type: {type: string}, document_date: {type: string, format: date}, currency_code: {type: string, minLength: 3, maxLength: 3}, remark: {type: string}, lines: {type: array, minItems: 1, items: {type: object, required: [description, quantity, unit_price], properties: {fee_type_id: {type: integer}, service_order_container_id: {type: integer}, description: {type: string}, quantity: {type: number}, unit_price: {type: number}, discount_amount: {type: number}, tax_rate: {type: number}}}}}
...
    AttachmentPresignRequest:
      type: object
      required: [file_name, mime_type, file_size_bytes, entity_type, entity_id]
      properties: {file_name: {type: string}, mime_type: {type: string}, file_size_bytes: {type: integer}, entity_type: {type: string}, entity_id: {type: integer}, attachment_role: {type: string}}
...
    ServiceOrder: {type: object, properties: {id: {type: integer}, service_order_no: {type: string}, organization_id: {type: integer}, branch_id: {type: integer}, status: {type: string}}}
    PagedServiceOrderResponse: {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/ServiceOrder'}}, meta: {$ref: '#/components/schemas/PageMeta'}}}
...